test: use a common create_topics interface
Centered around a TopicSpec object that stores various topic properties
and settings.
