+++ tests/rptest/tests/describe_topics_test.py
@@ -10,13 +10,15 @@
 from ducktape.mark import matrix
 from rptest.tests.redpanda_test import RedpandaTest
 
+from rptest.clients.types import TopicSpec
 from rptest.clients.kafka_cli_tools import KafkaCliTools
 
 
 class DescribeTopicsTest(RedpandaTest):
+    topics = (TopicSpec(partitions=2, replication_factor=3), )
+
     def test_describe_topics(self):
         tools = KafkaCliTools(self.redpanda)
-        tools.create_topic("topic", partitions=2, replication_factor=3)
         output = tools.describe_topics()
         assert "partition_count=2" in output
         assert "replication_factor=3" in output
